changeCodeInt

+++ src/Main.java
@@ -3,8 +3,8 @@ public class Main {
         BMIService service = new BMIService();
         double metre = 1.87; // рост
         double kilogram = 98; // вес
-        double index = service.calculate(metre,kilogram);
+        int index = (int) service.calculate(metre,kilogram);
 
-        System.out.println((int)index);
+        System.out.println(index);
     }
 }
